PP-11244 Stop using ChargeEntity for building JSON responses
The only place we still use ChargeEntity as an API response model is
for the GET `/v1/frontend/tokens/{chargeTokenId}` endpoint.

Instead use the `FrontendChargeResponse` model that we use for other
charge responses consumed by the card connector app.
