+++ .secrets.baseline
@@ -153,88 +153,81 @@
         "line_number": 787
       },
       {
-        "type": "Hex High Entropy String",
-        "filename": "openapi/connector_spec.yaml",
-        "hashed_secret": "48ca4a65970b54002939b02e8d8c1b0167de7e4b",
-        "is_verified": false,
-        "line_number": 3028
-      },
-      {
         "type": "Base64 High Entropy String",
         "filename": "openapi/connector_spec.yaml",
         "hashed_secret": "1af9a059666b095ca066171bfa93b84e895134d8",
         "is_verified": false,
-        "line_number": 3396
+        "line_number": 3027
       },
       {
         "type": "Hex High Entropy String",
         "filename": "openapi/connector_spec.yaml",
         "hashed_secret": "7aaebcf1d5950305be73fbdda70fa0c55c6c8542",
         "is_verified": false,
-        "line_number": 3742
+        "line_number": 3548
       },
       {
         "type": "Hex High Entropy String",
         "filename": "openapi/connector_spec.yaml",
         "hashed_secret": "f6e49af5bc530ae85699c4fb4dab97b5d7ea9fc0",
         "is_verified": false,
-        "line_number": 3791
+        "line_number": 3597
       },
       {
         "type": "Hex High Entropy String",
         "filename": "openapi/connector_spec.yaml",
         "hashed_secret": "b7b82fee3d8bb620476c30e98864525528a30dc5",
         "is_verified": false,
-        "line_number": 4396
+        "line_number": 4152
       },
       {
         "type": "Hex High Entropy String",
         "filename": "openapi/connector_spec.yaml",
         "hashed_secret": "4e2117d267bb6ba687414034ca6f7699e03d0098",
         "is_verified": false,
-        "line_number": 4457
+        "line_number": 4213
       },
       {
         "type": "Hex High Entropy String",
         "filename": "openapi/connector_spec.yaml",
         "hashed_secret": "a1f69fc9bf1984c89e57ee993c7392c3d702bb93",
         "is_verified": false,
-        "line_number": 4479
+        "line_number": 4235
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "openapi/connector_spec.yaml",
         "hashed_secret": "2b5620026862563e61e31e6cfbeb7551148c39bc",
         "is_verified": false,
-        "line_number": 4658
+        "line_number": 4414
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "openapi/connector_spec.yaml",
         "hashed_secret": "be54950d938040748a64e6cbbe239bb85ed6ab38",
         "is_verified": false,
-        "line_number": 4661
+        "line_number": 4417
       },
       {
         "type": "Hex High Entropy String",
         "filename": "openapi/connector_spec.yaml",
         "hashed_secret": "690d31e7e1b8e4a3c8f5e160d55c3ca4bf8c8ef8",
         "is_verified": false,
-        "line_number": 4664
+        "line_number": 4420
       },
       {
         "type": "Hex High Entropy String",
         "filename": "openapi/connector_spec.yaml",
         "hashed_secret": "40a714e4a70aa9b75e73dc27be3a978fa98ee4e3",
         "is_verified": false,
-        "line_number": 5254
+        "line_number": 4986
       },
       {
         "type": "Hex High Entropy String",
         "filename": "openapi/connector_spec.yaml",
         "hashed_secret": "ec969b7613112e3df2b11d8f64ce37f0d20bf29d",
         "is_verified": false,
-        "line_number": 5265
+        "line_number": 4997
       }
     ],
     "src/main/java/uk/gov/pay/connector/charge/model/ChargeResponse.java": [
@@ -244,15 +237,6 @@
         "hashed_secret": "423e3dd3782e3527b4f9af75dd9faf9ebbb7889c",
         "is_verified": false,
         "line_number": 60
-      }
-    ],
-    "src/main/java/uk/gov/pay/connector/charge/model/domain/ChargeEntity.java": [
-      {
-        "type": "Hex High Entropy String",
-        "filename": "src/main/java/uk/gov/pay/connector/charge/model/domain/ChargeEntity.java",
-        "hashed_secret": "48ca4a65970b54002939b02e8d8c1b0167de7e4b",
-        "is_verified": false,
-        "line_number": 201
       }
     ],
     "src/main/java/uk/gov/pay/connector/charge/resource/ChargesApiResource.java": [
@@ -1048,5 +1032,5 @@
       }
     ]
   },
-  "generated_at": "2023-07-03T14:21:19Z"
+  "generated_at": "2023-07-03T16:13:16Z"
 }
