+++ src/main/java/uk/gov/pay/connector/token/exception/TokenNotFoundException.java
@@ -1,0 +1,11 @@
+package uk.gov.pay.connector.token.exception;
+
+import javax.ws.rs.WebApplicationException;
+
+public class TokenNotFoundException extends WebApplicationException {
+
+    public TokenNotFoundException(String message) {
+        super(message);
+    }
+
+}
