+++ src/main/java/uk/gov/pay/connector/token/model/domain/TokenResponse.java
@@ -2,6 +2,8 @@ package uk.gov.pay.connector.token.model.domain;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
+import uk.gov.pay.connector.charge.model.ChargeResponse;
+import uk.gov.pay.connector.charge.model.FrontendChargeResponse;
 import uk.gov.pay.connector.charge.model.domain.ChargeEntity;
 
 import java.util.Objects;
@@ -14,9 +16,9 @@ public class TokenResponse {
 
     @JsonProperty("charge")
     @Schema(description = "The charge associated with the token")
-    private final ChargeEntity charge;
+    private final FrontendChargeResponse charge;
 
-    public TokenResponse(boolean used, ChargeEntity charge) {
+    public TokenResponse(boolean used, FrontendChargeResponse charge) {
         this.used = used;
         this.charge = Objects.requireNonNull(charge);
     }
